backend/requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# AI/ML
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0

# Financial APIs
yfinance>=0.2.18
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# CORS
fastapi-cors>=0.0.6 